pyproject.toml
[tool.poetry]
name = "nomad-django"
version = "0.1.0"
description = ""
authors = ["izunaaaaa <art970@naver.com>"]
readme = "README.md"
packages = [{include = "nomad_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
black = "^22.12.0"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
